Participant class and DAO

+++ JavaGuruChat/src/com/seventysevenagency/chat/dao/ConversationDAO.java
@@ -4,7 +4,7 @@ import com.seventysevenagency.chat.domain.Conversation;
 
 public interface ConversationDAO {
 	public void create(Conversation conversation) throws DAOException;
-	public void delete(Integer id)  throws DAOException;
-	public Conversation select(Integer id)  throws DAOException;
+	public void delete(Long id)  throws DAOException;
+	public Conversation select(Long id)  throws DAOException;
 	public void update(Conversation conversation) throws DAOException;
 }
